Move Go code to service directory to make it easier to maintain a cleaner Docker image.
