+++ Dockerfile_arm64
@@ -5,7 +5,7 @@ WORKDIR /service
 RUN apt clean && apt-get update && apt-get -y install alien
 # install dependencies
 
-COPY . .
+COPY service ./
 
 RUN ls /service
