+++ service/preprocessor/preprocessor.go
@@ -3,9 +3,9 @@ package preprocessor
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/pennsieve/processor-pre-external-files/logging"
-	"github.com/pennsieve/processor-pre-external-files/models"
-	"github.com/pennsieve/processor-pre-external-files/util"
+	"github.com/pennsieve/processor-pre-external-files/service/logging"
+	"github.com/pennsieve/processor-pre-external-files/service/models"
+	"github.com/pennsieve/processor-pre-external-files/service/util"
 	"io"
 	"log/slog"
 	"net/http"
